# docs/swagger.yaml
basePath: /gradiol-api/api
definitions:
...
  dto.CreateDocumentReq:
    properties:
      content:
        type: object
      diagram_type:
        enum:
        - flowchart
        - erd
        - usecase
        - sequence
        - mindmap
        - blank
        type: string
      project_id:
        type: string
      title:
        maxLength: 200
        type: string
      view:
        type: object
      workspace_id:
        type: string
    required:
    - diagram_type
    - workspace_id
    type: object
...
  dto.DocumentListItem:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      diagram_type:
        type: string
      id:
        type: string
      project_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
      workspace_id:
        type: string
    type: object
  dto.DocumentListResp:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DocumentListItem'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
  dto.DocumentResp:
    properties:
      content:
        type: object
      created_at:
        type: string
      created_by:
        type: string
      diagram_type:
        type: string
      id:
        type: string
      project_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
      view:
        type: object
      workspace_id:
        type: string
    type: object
...
  dto.PaginationMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateDocumentReq:
    properties:
      content:
        type: object
      project_id:
        description: nullable — can move or set to null
        type: string
      title:
        maxLength: 200
        type: string
      view:
        type: object
    type: object
...
  model.Document:
    properties:
      content:
        items:
          type: integer
        type: array
      created_at:
        type: string
      created_by:
        type: string
      diagram_type:
        type: string
      id:
        type: string
      project_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
      view:
        items:
          type: integer
        type: array
      workspace_id:
        type: string
    type: object
  model.DocumentVersion:
    properties:
      content:
        items:
          type: integer
        type: array
      created_at:
        type: string
      created_by:
        type: string
      document_id:
        type: string
      id:
        type: string
      version:
        type: integer
      view:
        items:
          type: integer
        type: array
    type: object
...
  pkg.AppError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
host: asia-southeast1-renzip-478811.cloudfunctions.net
info:
  contact: {}
  description: Backend API for Graphic Diagram Online (GraDiOl)
  title: GraDiOl API
  version: "1.0"
paths:
...
  /documents:
    post:
      consumes:
      - application/json
      description: Creates a new document in a workspace/project
      parameters:
      - description: Document Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDocumentReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DocumentResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Create a new document
      tags:
      - documents
  /documents/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a document by ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Delete a document
      tags:
      - documents
    get:
      consumes:
      - application/json
      description: Retrieves the full content and metadata of a specific document
        by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Get document detail
      tags:
      - documents
    put:
      consumes:
      - application/json
      description: Updates a document's content and metadata
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Document Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDocumentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Update a document
      tags:
      - documents
  /documents/{id}/versions:
    get:
      consumes:
      - application/json
      description: Retrieves the history of saved versions for a document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DocumentVersion'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Get document versions
      tags:
      - documents
  /documents/{id}/versions/{version}/restore:
    post:
      consumes:
      - application/json
      description: Restores a document to a specific previous version
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pkg.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Restore document version
      tags:
      - documents
  /documents/recent:
    get:
      consumes:
      - application/json
      description: Retrieves the user's recently accessed/updated documents
      parameters:
      - description: Number of documents to retrieve
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Document'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Get recent documents
      tags:
      - documents
...
  /projects/{id}/documents:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of documents inside a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by diagram type
        in: query
        name: diagram_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.AppError'
      security:
      - BearerAuth: []
      summary: Get all documents by project
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"